+++ test/json_rpc/manager.py
